se restructura la carpeta pages globales para que tenga la vista basica sin registrarse
