+++ src/front/js/layout.js
@@ -16,9 +16,9 @@ import { Datos_contrato } from "./pages/datos_contrato";
 import { Ordenes } from "./pages/ordenes";
 import { Datos_orden } from "./pages/datos_orden";
 import { Cuatrocerocuatro } from "./pages/pages_globales/404";
-import { Contactenos } from "./pages/contactanos";
+import { Contactenos } from "./pages/pages_globales/contactanos";
 import { RestablecerContrasena } from "./pages/restablecer_contrasena";
-import { CambiarContrasena } from "./pages/cambiar_contrasena";
+import { CambiarContrasena } from "./pages/pages_globales/cambiar_contrasena";
 import { VerFotosOrden } from "./pages/ver_fotos_orden";
 import { Listausuarios } from "./pages/listausers";
 import { Datosuser } from "./pages/Datosuser";
