+++ src/front/js/pages/pages_globales/contactanos.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect, useContext } from "react";
 import { Link, useParams } from "react-router-dom";
-import { Context } from "../store/appContext";
+import { Context } from "../../store/appContext";
 //import figure from "react-bootstrap/figure";
-import pedroyanezIMG from "../../img/pedroyanez.jpg";
-import luissalgado from "../../img/luissalgado.jpg";
-import felipevalenzuela from "../../img/felipevalenzuela.jpeg";
-import "../../styles/app.scss";
-import "../../styles/contactenos.scss";
+import pedroyanezIMG from "../../../img/pedroyanez.jpg";
+import luissalgado from "../../../img/luissalgado.jpg";
+import felipevalenzuela from "../../../img/felipevalenzuela.jpeg";
+import "../../../styles/app.scss";
+import "../../../styles/contactenos.scss";
 
 export const Contactenos = props => {
 	return (
